refactor: 폴더 이름 clip 으로 변경
