+++ src/entities/clip/index.ts
