+++ src/entities/clip/mock/index.ts
